+++ nzrealme/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = 'Darryl Cousins'
 __email__ = 'darryljcousins@gmail.com'
-__version__ = '0.1.0'
+__version__ = '0.0.1'