+++ HISTORY.rst
@@ -3,7 +3,7 @@
 History
 -------
 
-0.1.0 (2014-01-11)
----------------------
+Development (2014-01-11)
+------------------------
 
-* First release on PyPI.
+* continuing development